+++ packages/in_app_messaging/lib/src/presentation/presenter/presenter.dart
@@ -1,3 +1,4 @@
 export 'dynamic_message_handle_factory.dart';
 export 'dynamic_message_handle.dart';
 export 'dynamic_message_presenter.dart';
+export 'presentation_outcome.dart';